requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
boto3==1.34.0
minio==7.2.7
Pillow==10.3.0
celery==5.3.6
redis==5.0.4
python-magic==0.4.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
aiofiles==23.2.1